XQuartz: pbproxy: Fix NSObject memory leaks by properly using the NSAutoreleasePool.   Now the usage is consistent.  In x_input_run() we create a pool, and release it after processing the XEvents.
Add some getpid() output to main for debugging.  It needs a bit more
testing before the next release.

Don't retain the NSPasteboard as the old code did.  That may have
contributed to the leak, and it made it so that we needed the
NSAutoreleasePool created in main().

Remove the _known_types, and _pasteboard instance variables from
the x_selection class.  They aren't needed anymore.

The leaks program now indicates 0 leaks after some usage.  I want
to test further, but this seems much better, and my memory usage
graph indicates it's not growing.