XQuartz: GL: Convert some of the indirect.c code to work with the change to __GLXconfig.
Fix some obvious errors with uninitialized new members of GLX structs.

Add a type argument for the screen's drawable creation function, and pass the type to
__glXDrawableInit().

GLX still doesn't work.  There is more work needed to get it working again.  This gets
it building at least, although there are some warnings in the dispatch table setup,
that seem to be caused by const differences.

I also #if 0ed a bunch of function bodies that will need to be revisited soon.