XQuartz: GL: Add 1 __GLXconfig for testing purposes.  Now glxgears runs.
It looks much better with the new GLX/libGL, perhaps because of the old libGL
not understanding all of the tags written.  At least glxgears works with both.

glxgears is not a test for all OpenGL features, so most likely some things
are broken still.  Direct rendering is known to not work, or at least being
reported as not working.