XQuartz: GL: Add a branch for a NULL DrawablePtr.
This may fix a bug where a NULL structure pointer was accessed resulting in:
KERN_PROTECTION_FAILURE at 0x0000000000000004

I studied the structures in use, and I believe it was the pDraw->id that resulted in
this fault in attach().

The layout of the _Drawable structure is 4 chars, and then the id, so the offset
seems correct for such a bug.

Also make it return an error for now, and ErrorF() to help us figure out where it's
going wrong elsewhere to cause a NULL DrawablePtr.

I also made various formatting changes, because the formatting was inconsistent.
Next time I will probably do the formatting in a separate commit, but hindsight is
20/20.

I will apply a fix to the 1.5 branch after this, in a similar manner.