XQuartz: GL: Fix a problem with uninitialized data in __glXAquaScreenProbe.
I added a memset to fix the issue with an uninitialized function pointer,
and potentially other GLX data in ->base.

Also use xalloc and xfree, instead of malloc.

Enable the alpha bits from the __GLcontextModes.  The alphaBits are no
longer always 0 as the previous comment eluded to.

Add some commentary about how some of the data structures work.

Cleanup some of the GLAQUA_DEBUG_MSG.

This was tested with various Mesa demos in a variety of situations over
indirect.