XQuartz: Fix alpha to be 1 (screenshots bug, etc)
This is related to: bc964ff1e3f3d8ab590eab40d767c6241e9480d4
XQuartz: Stab at fixing the alpha 0/1 bug (screenshots, etc) by pulling in some old code that got gutted from rootless.

which was on the 1.4 branch and implemented in fbPaintWindow.  Now that fbPaintWindow is gone, this is now in miPaintWindow().