Block signals while delivering XTest events (fixes bug 23100)
XTest uses the same event list as all input devices, and input devices
use it at SIGIO time. This causes mass confusion when a SIGIO event
occurs in the middle of XTest event processing.

The master branch has a much different fix, using a separate input
queue which reduces latency that may be caused by XTest event queuing;
it's a 'nicer' fix in some ways, but this simple fix should solve the
problem.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Keith Packard <keithp@keithp.com>