Revert "os: Prevent backtrace from being stopped in noreturn functions."
This reverts commit 579715f830fbbca9e1ecb17dc18176132f5969e7.

The patch is not needed anymore. I haven't encountered backtrace
problems with GCC 4.3.3. Even if the problems still persisted, this
commit should be removed and instead the definition of _X_NORETURN
should be modified to be empty if GCC/ARM is used. However, currently
it seems that ARM backtraces are OK even if _X_NORETURN is used and
-mapcs-frame is not defined in CFLAGS.

Signed-off-by: Rami Ylimäki <rami.ylimaki@vincit.fi>
Reviewed-by: Jamey Sharp <jamey@minilop.net>
(cherry picked from commit 2b364bf970b2ce6829af656990c33afd0d365f3c)