dix: only transform valuators when we need them.
Unconditionally drop the valuators back into the mask when they were there
in the first place. Otherwise, sending identical coordinates from the driver
on a translated device causes the valuator mask to be alternatively
overwritten with the translated value or left as-is. This leads to the
device jumping around between the translated and the original position.

The same could be achieved with a valuator_mask_unset() combination.

Testcase:
xsetwacom set "device name" MapToOutput VGA1
Then press a button on the device, cursor jumps between the two positions.

Introduced in 31737fff08ec19b394837341d5e358ec401f5cd8

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Daniel Stone <daniel@fooishbar.org>
Reviewed-by: Chase Douglas <chase.douglas@canonical.com>